Extract SendAppStoreReceiptOperation
